+++ src/Spago/CLI.hs
@@ -1,6 +1,6 @@
 module Spago.CLI
   ( module Spago.CLI
-  , CLI.optionsExt
+  , CLI.options
   ) where
 
 import Spago.Prelude
@@ -21,9 +21,12 @@ import Spago.Version (VersionBump (..))
 
 -- | Commands that this program handles
 data Command
+
+  -- | Default catch-all command
+  = Default ShowVersion
   
   -- | Initialize a new project
-  = Init Force TemplateComments
+  | Init Force TemplateComments
 
   -- | Install (download) dependencies defined in spago.dhall
   | Install [PackageName]
@@ -109,6 +112,7 @@ parser = do
     <|> publishCommands
     <|> otherCommands
     <|> oldCommands
+    <|> ( Default <$> versionOpt )
   pure (command, opts)
   where
     cacheFlag =
@@ -161,6 +165,12 @@ parser = do
     pursArgs        = many $ CLI.opt (Just . PursArg) "purs-args" 'u' "Arguments to pass to purs compile. Wrap in quotes."
     buildOptions  = BuildOptions <$> watch <*> clearScreen <*> allowIgnored <*> sourcePaths <*> srcMapFlag <*> noInstall
                     <*> pursArgs <*> depsOnly <*> beforeCommands <*> thenCommands <*> elseCommands
+
+    -- Opts.flag' creates a parser with no default value. This is intended.
+    -- We want this parser to fail if it does not get a --version flag, rather
+    -- than defaulting to NoShowVersion.  We rely on the parser failure to show
+    -- the help message.
+    versionOpt  = Opts.flag' DoShowVersion (Opts.long "version" <> Opts.help "Show spago version")
 
     -- Note: by default we limit concurrency to 20
     globalOptions = GlobalOptions <$> quiet <*> verbose <*> veryVerbose <*> (not <$> noColor) <*> usePsa